pyproject.toml
[project]
name = "lastapi"
version = "0.1.0"
description = "An example FastAPI application dedicated to the interview"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi",
    "httpx",
    "uvicorn[standard]",
    "pydantic",
    "python-dotenv",
    "pytest-asyncio"
]

[project.scripts]
main = "app.main:main"